게시글 유저 연동

+++ src/main/java/com/example/demo/controller/ArticleController.java
@@ -8,10 +8,13 @@ import java.util.List;
 import java.util.Optional;
 
 import com.example.demo.dao.ArticleDAO;
+import com.example.demo.dao.UserDAO;
 import com.example.demo.dto.Article;
+import com.example.demo.dto.User;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -29,13 +32,15 @@ import org.springframework.web.bind.annotation.RequestBody;
 public class ArticleController {
     @Autowired
     ArticleDAO articleDAO;
+    @Autowired
+    UserDAO userDao;
 
     private final Log logger = LogFactory.getLog(getClass());
 
     @GetMapping(value = "/article/all")
     public Object GetAllArticle() {
-        List<Article> article = articleDAO.findAll();
-        return article;
+        List<Article> articles = articleDAO.findAll();
+        return articles;
     }
 
 
@@ -48,9 +53,10 @@ public class ArticleController {
     @PostMapping(value="/article")
     public Object PostArticle(@RequestBody Article article) {
         // 유저정보 가져오기
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        logger.info(auth);
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userDao.findById(username).get();
         article.setCreateDate(LocalDateTime.now());
+        article.setUser(user);
         return articleDAO.save(article);
     }
 
@@ -58,18 +64,33 @@ public class ArticleController {
     public Object PutArticle(@PathVariable("id") Long id, @RequestBody Article newArticle) {
 
         Optional<Article> articles = articleDAO.findById(id);
-        return articles.map(
-            article -> {
-                article.setTitle(newArticle.getTitle());
-                article.setContent(newArticle.getContent());
-                return articleDAO.save(article);
-            }
-        );
+        if(!articles.isPresent()){
+            return new ResponseEntity<>("없는 게시글", HttpStatus.BAD_REQUEST);
+        }
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!articles.get().getUser().equals(username)){
+            return new ResponseEntity<>("게시글 작성자가 아님", HttpStatus.FORBIDDEN);
+        }
+        Article article = articles.get();
+        article.setTitle(newArticle.getTitle());
+        article.setContent(newArticle.getContent());
+        articleDAO.save(article);
+        return new ResponseEntity<>("수정완료", HttpStatus.OK);
+
     }
 
     @DeleteMapping(value = "/article/{id}")
-    public void DeleteArticle(@PathVariable("id") Long id) {
+    public Object DeleteArticle(@PathVariable("id") Long id) {
+        Optional<Article> articles = articleDAO.findById(id);
+        if(!articles.isPresent()){
+            return new ResponseEntity<>("없는 게시글", HttpStatus.BAD_REQUEST);
+        }
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        if(!articles.get().getUser().equals(username)){
+            return new ResponseEntity<>("게시글 작성자가 아님", HttpStatus.FORBIDDEN);
+        }
         articleDAO.deleteById(id);
+        return new ResponseEntity<>("삭제완료", HttpStatus.OK);
     }
     
 }
